refactor: generalize relative time, use seconds instead of minutes
