+++ src/popup/locales/en.json
@@ -1026,6 +1026,7 @@
     ]
   },
   "redirecting": "Redirecting",
+  "seconds": "{n} seconds | {n} second | {n} seconds",
   "minutes": "{n} minutes | {n} minute | {n} minutes",
   "hours": "{n} hours | {n} hour | {n} hours",
   "days": "{n} days | {n} day | {n} days"
