+++ src/popup/router/components/FormatDate.vue
@@ -5,6 +5,8 @@
 </template>
 
 <script>
+import { secondsToRelativeTime } from '../../../filters/toRelativeTime';
+
 export default {
   props: {
     date: { type: Date, required: true },
@@ -12,9 +14,11 @@ export default {
   computed: {
     formatDate() {
       const today = new Date();
-      const tipDate = this.date;
-      const isToday = today.toDateString() === tipDate.toDateString();
-      return isToday ? tipDate.toLocaleTimeString('en-GB') : tipDate.toLocaleDateString('en-US');
+      const { date } = this;
+      const lessThanFourDaysAgo = today - date < 4 * 24 * 60 * 60 * 1000;
+      return lessThanFourDaysAgo
+        ? `${secondsToRelativeTime((today - date) / 1000)} ago`
+        : date.toLocaleDateString('en-US');
     },
     wholeDateAndTime() {
       return this.date.toLocaleString('en-US', { hourCycle: 'h24' });
